angel-ultracon/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.4
redis==5.0.4
nats-py==2.6.0
PyNaCl==1.5.0
orjson==3.10.3
jsonschema==4.23.0
